Handle some XdmValue conversions
Implement some conversion in the {@link Convert} class. The saxDestination conversion loses type information in some cases, e.g. fn-transform-63 from XQTS has a <xsl:template name='main' as='xs:integer’>. which alongside "delivery-format":"raw" fails to deliver an int.

We implement a vestigial Convert class to use the XdmValue generated by the alternative form of call to the transformer with a template, in this case.